Fix an off-by-one error demangling satellite assembly names
When assemblies are packaged in the APK archive, their names are
"mangled" so that we don't create clashes with "real" shared libraries,
which are placed in the same directory.  Additionally, since the `lib/`
directory in the APK archive supports only single level of directories,
we must encode satellite assembly culture in the file name, instead of
putting it in a subdirectory.

Unfortunately, when the code was committed, an off-by-one error crept
in, when converting satellite assembly names to their `culture/name`
for.  As the result, the first character of the culture was omitted
which would cause the satellite assemblies not to be loaded properly.

Fix the issue by using the correct mangled name prefix length.