[Mono.Android] Add documented AndroidManifest.xml-related properties (#9096)
Fixes: https://github.com/dotnet/android/issues/9092

Context: 3ab74db2c669fc234c6d60ed9884aecec8dfe61a

Previously in commit 3ab74db2, we created a process for automatically
determining the possible attributes for elements in
`AndroidManifest.xml` and generating the `[Application]`, `[Activity]`,
etc. attributes within `Mono.Android.dll` for users to consume.  At the
time, we merely created the process, but did not use it to expose any
new attributes that have been added in previous Android API levels that
we never exposed.

This commit exposes all Google-documented manifest attributes for the
elements we support.  The undocumented manifest attributes were moved
around in the file to a section denoted as "undocumented".