Fix an issue with incremental builds.
We have been getting on and off reports of the following
errors in Android for quite a while now.

```
Resources/values/styles.xml(2): error APT2260: resource attr/colorPrimary (aka Microsoft.Maui:attr/colorPrimary) not found.
```

It was always very confusing as to why this was happening.
No one ever seemed to be able to figure out why these files
were deleted.

Well it turns out they were not deleted, just not included.
While building maui I came across this issue and managed to
capture a `.binlog`.

It turns out that the `_CleanMonoAndroidIntermediateDir` target
was running. And it was deleting certain files. This target was
running because `_CleanIntermediateIfNeeded` was running.
That was running because the `build.props` file had changed...
That was because the NuGet `project.assets.json` file had a new
timestamp! On an incremental build....

So it looks like NuGet sometimes touches the `project.assets.json`
file even if it does not change. We can't blame them for that as
we do similar things.

The next confusing thing and the principal cause is that the
`libraryprojectimports.cache` was being deleted, probably buy the
call to `_CleanMonoAndroidIntermediateDir`. However the
`_ResolveLibraryProjectImports.stamp` file was left in place.
This means the `_ResolveLibraryProjectImports` does NOT run.
So we end up NOT including ANY of the resource `res.zip` files when
calling `aapt2`. The `_ResolveLibraryProjectImports` target did not
include `libraryprojectimports.cache` in its `Outputs`. So if the file
did not exist, but the stamp file did the target would not run.
It is confusing because in the `_CleanMonoAndroidIntermediateDir` we delete
the `libraryprojectimports.cache` and the entire `$(IntermediateOutputPath)stamp`
directory..

So lets include `libraryprojectimports.cache` in the outputs for starters.
Then update the `_CreatePropertiesCache` target to use a hash of the
`project.assets.json` file rather than a timestamp. This way we really
only run the build if something actually changes.