[Xamarin.Android.Build.Tasks] fix `$(DestinationSubDirectory)` double slash
I noticed in a `.binlog` in .NET 9:

    ProcessAssemblies
        ...
        OutputItems
            ...
            D:\.nuget\packages\microsoft.maui.controls.core\9.0.0-preview.7.24407.4\lib\net9.0-android35.0\ar\Microsoft.Maui.Controls.resources.dll
                DestinationSubDirectory = arm64-v8a\ar\\

I believe this causes an issue, later down the line:

    FastDeploy
    ...
    Remove redundant file files/.__override__/arm64-v8a/ar/Microsoft.Maui.Controls.resources.dll

So, `<FastDeploy/>` seems to delete every `Microsoft.Maui.Controls.resources.dll`
file, because it is passed in:

    obj\Debug\net9.0-android\android\assets\arm64-v8a\ar\Microsoft.Maui.Controls.resources.dll
    TargetPath = arm64-v8a\ar\\Microsoft.Maui.Controls.resources.dll

I think this was introduced in 86260ed3, we can fix it by only
appending to `%(DestinationSubDirectory)` if it *doesn't* end with a
`DirectorySeparatorChar`.

There may be additional safety we could add to the `<FastDeploy/>` task,
in a separate PR.

To test this going forward, I'm going to review our MSBuild performance
tests, and see if we can adjust the timings there.