android-sdk-installer version harmonization
Context: https://aka.ms/AndroidManifestFeed/d17-12

The "Xamarin" manifest used the `InstallAndroidDependencies` task
from android-sdk-installer, and contains only a select number of
Android SDK and JDK package versions.

Until android-sdk-installer provides a `Xamarin.Installer.Common.props`,
the versions we use need to be in harmony with the Xamarin manifest,
so that the `InstallAndroidDependencies` tests can pass.

Partially revert 0aede13b and provision JDK 17.0.8 instead of 17.0.11.

Explicitly set `$(AndroidNdkVersion)`=26.3.11579264, as that's the
latest version in the manifest.