[Xamarin.Android.Build.Tasks] Rethink default property values (#9155)
Context: 5d1ac0a5dfa12e92fc45b384237e929576a418e4
Context: https://github.com/dotnet/android-tools/commit/a6a23bb5c59be8164e115db9ddf750c583c7b3c9
Context: https://learn.microsoft.com/en-us/visualstudio/mac/what-happened-to-vs-for-mac?view=vsmac-2022

Commit 5d1ac0a5 had a brilliant idea:

> Of particular note is xamarin/xamarin-android-tools@a6a23bb:
> [ed note: now dotnet/android-tools@a6a23bb]
>
> In order to ensure that both Xamarin.Android and the macOS installer
> use the same default version numbers for Android SDK components, we are
> now using `Xamarin.Android.Tools.Versions.props` to contain default
> version number information.  This file is maintained in the
> xamarin/xamarin-android-tools repo and will be imported via
> `external/xamarin-android-tools`:

In the intervening four years (!), we've learned something, and the
environment has changed.

What we've learned is that this approach is "bump-heavy": in order to
update the default value of e.g. `$(AndroidSdkPlatformVersion)`,
we'd need to update:

 1. dotnet/android-tools
 2. xamarin/androidtools
 3. Plus other repos that depend on (1) or (2) that are used by (4)
 4. xamarin/monodroid
 5. dotnet/android

This is, to put it mildly, cumbersome.

Additionally, it's no longer necessary: Visual Studio for Mac will be
retired on 2024-Aug-31, so there is no longer a need to have a common
file shared between the .NET for Android SDK and Visual Studio for Mac.

Instead of considering `Xamarin.Android.Tools.Versions.props` as the
canonical source of truth:

 1. Update `Xamarin.Android.Common.targets` to now optionally include
    a new `Xamarin.Installer.Common.props` file, if present.
    This will allow e.g. xamarin/android-sdk-installer to eventually
    provide default property values which are tied to the in-use
    "Xamarin Manifest" for package installation, e.g.
    `$(AndroidCommandLineToolsVersion)`, `$(JavaSdkVersion)`.

 2. Update `Xamarin.Android.Common.props.in` so that the following
    properties are set *before* importing
    `Xamarin.Android.Tools.Versions.props`:

      * `$(AndroidCommandLineToolsVersion)`
      * `$(AndroidNdkVersion)`
      * `$(AndroidSdkBuildToolsVersion)`
      * `$(AndroidSdkEmulatorVersion)`
      * `$(AndroidSdkPlatformVersion)`
      * `$(AndroidSdkPlatformToolsVersion)`
      * `$(JavaSdkVersion)`

    This will allow us to update these values when this repo ads
    support for these versions.

Finally, the ["Xamarin" manifest][0] is used by the
`InstallAndroidDependencies` target from xamarin/android-sdk-installer,
and contains only a select number of Android SDK and JDK versions.

Until android-sdk-installer provides `Xamarin.Installer.Common.props`,
the versions we use need to be in harmony with the Xamarin manifest,
so that the `InstallAndroidDependencies` tests can pass.

Explicitly set `$(AndroidNdkVersion)`=26.3.11579264, as that's the
latest version in the manifest.

[0]: https://aka.ms/AndroidManifestFeed/d17-12