[Mono.Android] Use .NET version of mdoc (#9225)
Fixes: https://github.com/dotnet/android/issues/7974

We've been intermittently running into issues when building Mono.Android
on our PR build pool ever since commit 41b0124f which made mono an
optional build dependency.  Update mdoc invocations to use the .NET
version of the tool to remove one of the last parts of the build that
still depends on mono.