[Xamarin.Android.Build.Tasks] Fix an issue with incremental builds (#9183)
We have been getting on and off reports of the following build errors
in .NET for Android for quite a while now:

	Resources/values/styles.xml(2): error APT2260: resource attr/colorPrimary (aka Microsoft.Maui:attr/colorPrimary) not found.

It was always very confusing as to why this was happening.  No one
ever seemed to be able to figure out why these files were deleted.

Well it turns out they were not deleted, just not included. 
While building dotnet/maui I came across this issue and managed to
capture a `.binlog`.

It turns out that the `_CleanMonoAndroidIntermediateDir` target was
running, and it was deleting certain files.  This target was running
because `_CleanIntermediateIfNeeded` was running.  That was running
because the `build.props` file had changed, which was because the
NuGet `project.assets.json` file had a new timestamp!

On an incremental build!

So it looks like NuGet sometimes touches the `project.assets.json`
file even if it does not change.  We can't blame them for that as we
do similar things.

The next confusing thing and the principal cause is that the
`libraryprojectimports.cache` was not present.  However, the
`_ResolveLibraryProjectImports.stamp` file *was* present.  This means
that the `_ResolveLibraryProjectImports` target does NOT run, so we
end up NOT including ANY of the resource `res.zip` files when calling
`aapt2`.  The `_ResolveLibraryProjectImports` target did not include
`libraryprojectimports.cache` in its `Outputs`, so if the file did
not exist, but the stamp file *did*, then the target would not run.

It is confusing because in the `_CleanMonoAndroidIntermediateDir` we
delete the `libraryprojectimports.cache` and the entire
`$(IntermediateOutputPath)stamp` directory.

After much searching I did something while maui was opened in VSCode:
I deleted its `artifacts` directory to clean up and start a new build
and ..... it magically re-appeared!!!

Then I thought... Design Time Builds!!!!

So the problem we have is this: in our current system the
Design Time Builds and the main builds all use the same stamp files!!!
What was happening is a design time build was running before the main
build, and it was creating a design time cache file into
`$(_AndroidIntermediateDesignTimeBuildDirectory)`
(`$(IntermediateOutputPath)designtime`), but place the stamp files
INTO THE SAME LOCATION as the main build would.  As a result the main
build would skip the `_ResolveLibraryProjectImports` target
COMPLETELY even if the output cache file was not present.  Once that
happens it will NEVER recover until you delete the bin/obj directories.

This only happened occasionally because it would depend on if you
have the project open in a IDE and when the IDE decided to run a
Design Time Build.

Fix this by giving design-time builds their own stamp directory.
It might cause some targets to run during a design time build but I
think that is preferred to a failing main build.

To work around the NuGet causing a build when it touches a the
`project.assets.json` file we now use `<GetFileHash/>`.  This means
we only refresh the library projects if the contents of the
`project.assets.json` file change, rather than the timestamp.

Finally, there was a bug where we would automatically fallback to the
NuGet lock file even if we found the `project.assets.json` file.