[Xamarin.Android.Build.Tasks] Reduce BuildDependsOn duplication
I was looking to extend the BuildDependsOn target ordering used to
structure application and library builds and noticed a fair amount of
duplication that may benefit from some consolidation.

A new `AndroidBuildDependsOn` property has been added to include the
core targets that all builds depend on.

The  `_CheckNonIdealBindingConfigurations` target has been removed and
merged with `_CheckNonIdealConfigurations`.

The `_CheckProjectItems` target will now run for both app and library
projects, as the corresponding `CheckProjectItems` task already accepts
`$(AndroidApplication)` as a parameter.

The `_AddAndroidDefines` target has been removed in favor of the more
widely used `AndroidPrepareForBuild` target, both of which are empty
targets that depend on `$(_OnResolveMonoAndroidSdks)`.

The `_CalculateAssetsWithAssetPackMetaData` target has been removed, and
the library error condition has been moved into `_CalculateAssetPacks`.

The `_RemoveLegacyDesigner` target has been removed from the library
specific target list, it will run as part of the `UpdateGeneratedFiles`
which is included in `CoreCompileDependsOn`.