[tests] Add a Performance Test for FastDeploy Task (#9258)
We need to check the speed of the `<FastDeploy/>` Task to make sure
we do not regress.  This will also help us keep track of any
improvements we make. 

The timeout has been set to 20 seconds for an initial deploy on an
emulator to take into account the fact that the emulators are really
slow on CI.