[trimming] `TrimMode=full` in debug mode, should enable analyzers
For AOT scenarios, projects would set:

    <PublishAot>true</PublishAot>

For both `Debug` and `Release`, this allows you to get the same set of
analyzers in both configurations. You'd want to see the same warnings
for both.

For Android, the project template is currently doing:

    <PropertyGroup Condition="'$(Configuration)' == 'Release'">
        <TrimMode>Full</TrimMode>
    </PropertyGroup>

But this would result in a *different* set of warnings between `Debug`
and `Release` mode!

Instead, we can do:

    <PropertyGroup>
        <TrimMode>Full</TrimMode>
    </PropertyGroup>

And then add a new default, such as:

    <EnableTrimAnalyzer Condition="'$(EnableTrimAnalyzer)' == '' and '$(TrimMode)' == 'full'">true</EnableTrimAnalyzer>

`TrimMode=Full` does *not* enable the trimmer, so other defaults in
`Debug` should remain unchanged.

So, the new behavior is:

    * `Configuration=Debug`
    * `PublishTrimmed=false` (default, no trimmer)
    * `TrimMode=Full` (project template)
    * `EnableTrimAnalyzer=true` (new default)
    * You get the same warnings in `Debug` and `Release` mode.

I also reworded the commend in the project template slightly, to
mention it enables analyzers.