[XABT] Update AndroidGradleProject Bind and Pack defaults
Context: https://github.com/dotnet/android/commit/dbdc5fadca60b79518739fea49bdfee58a748bf0

Updates the default values of the `%(Bind)` and `%(Pack)` metadata
forwarded to the created `@(AndroidLibrary)` item to better match the
defaults on that item. Leaving these empty did not allow the defaults
declared in the `AndroidLibrary` ItemDefinitionGroup to be used as
expected in commit dbdc5fa.