[Xamarin.Android.Build.Tasks] AndroidGradleProject Bind,Pack=True (#9341)
Context: dbdc5fadca60b79518739fea49bdfee58a748bf0

Updates the default values of the `%(Bind)` and `%(Pack)` metadata
forwarded to the created `@(AndroidLibrary)` item to better match the
defaults on that item (True, by default).  Leaving these empty did
not allow the defaults declared in the `@(AndroidLibrary)`
ItemDefinitionGroup to be used as expected in commit dbdc5fa.

This allows the compiled output of `@(AndroidGradleProject)` to be
bound, making it usable to the containing C# project.