[Xamarin.Android.Build.Tasks] Add Inputs for _GenerateJavaStubs (#9319)
Context: https://github.com/xamarin/monodroid/pull/1547

This commit updates the Inputs for the `_GenerateJavaStubs` and
`_GeneratePackageManagerJava` targets so that the
`@(AndroidEnvironment)` and `@(LibraryEnvironments)` items are only
included when we are building a release build, or a build which
embeds the assemblies into the apk.

This is so that we can eventually fast deploy these files to the
device; see also xamarin/monodroid#1547. 

To *not* attempt to (eventually) fast deploy `@(AndroidEnvironment)`
and `@(LibraryEnvironments)`, set the (*internal*!) MSBuild property
`$(_AndroidFastDeployEnvironmentFiles)` to False.

A new unit test has been added to make sure that the environment
values can be used correctly on device.