[Xamarin.Android.Build.Tasks] move `Console.WriteLine()` messages to MSBuild log (#9359)
When running builds at the command-line (with the default, new
"MSBuild Terminal Logger"), `Console.WriteLine()` output can make the
text not look so great:

	> dotnet build -c Release -bl
	Restore complete (0.6s)
	You are using a preview version of .NET. See: https://aka.ms/dotnet-support-policy
	hellomaui                                                                                      _BuildApkEmbed (18.9s)
	Number of descriptors: 116; index entries: 232
	hellomaui succeeded (33.3s) → bin\Release\net9.0-android\android-arm64\hellomaui.dll
	Build succeeded in 46.8s

The message `Number of descriptors: 116; index entries: 232` appears,
but so does another one.  You can watch the terminal logger clear and
rewrite messages, which is not ideal.

Let's move these to the MSBuild logger, so they will show up in
`.binlog` files, but not in the console output.