Move all of assembly store building code to a separate class (#9382)
Simplify the task of constructing assembly stores by moving all the code
which actually stashes assemblies (and their config and debug files, if found
and required) in the assembly store to a separate utility class, `AssemblyStoreBuilder`.  

Additionally, move a portion of assembly packaging code from `BuildApk` to a helper class.

This makes it easier to build the stores outside of the `BuildApk` task and further 
simplifies `BuildApk` code.