[tests] use x86_64 API 23 image on nightly pipeline
Context: https://github.com/dotnet/runtime/issues/106025
Context: https://discord.com/channels/732297728826277939/732297837953679412/1296764055654240317

`Mono.Android-Tests` on API 23 crash on launch with:

    09-17 14:41:32.839  2758  2758 I monodroid-assembly: Failed to load shared library '/data/app/Mono.Android.NET_Tests-1/split_config.x86.apk!/lib/x86/libSystem.Native.so'. dlopen failed: cannot find "libc.so" from verneed[1] in DT_NEEDED list for "/data/app/Mono.Android.NET_Tests-1/split_config.x86.apk"

But this same library appears to be correct, and loads on all of:

* x86 API 21
* x86_64 API 23
* x86 API 24

This leads us to believe there is some kind of bug with loading native
libraries on the API 23 x86 emulator image.

For now, let's just target x86_64 on API 23 for the time being.